releases fix

+++ reporter/MessageViewController.m
@@ -168,7 +168,7 @@
         if (cell == nil) {
             cell = [[[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:CellIdentifier] autorelease];
         }
-        UIImprovedImage *image = [message.photos objectAtIndex:indexPath.row-1];
+        UIImprovedImage *image = [message.photos objectAtIndex:(NSUInteger)(indexPath.row-1)];
         cell.imageView.image = image;
         cell.textLabel.text = [NSString stringWithFormat:@"%d",indexPath.row];
         
@@ -188,7 +188,7 @@
 - (void)tableView:(UITableView *)tableView commitEditingStyle:(UITableViewCellEditingStyle)editingStyle forRowAtIndexPath:(NSIndexPath *)indexPath
 {
     if (editingStyle == UITableViewCellEditingStyleDelete) {
-        [message.photos removeObjectAtIndex:indexPath.row-1];
+        [message.photos removeObjectAtIndex:(NSUInteger)(indexPath.row-1)];
         [tableView deleteRowsAtIndexPaths:[NSArray arrayWithObject:indexPath] withRowAnimation:UITableViewRowAnimationFade];
     }
 }
